requirements.txt
altgraph==0.17.4
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Dance==7.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
markdown2==2.4.12
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
pefile==2023.2.7
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==69.0.3
SQLAlchemy==2.0.23
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
URLObject==2.4.3
Werkzeug==3.0.1
